test(cli): share resource command harness
