+++ crates/cli/tests/support/mod.rs
@@ -1,0 +1,1 @@
+pub(crate) mod resource_cli;
